feat: add a check

+++ ecom_backend/src/main/java/com/example/ecom_backend/services/ProductService.java
@@ -29,6 +29,9 @@ public class ProductService {
     }
 
     public void deleteById(Long id) {
+        if (!repo.existsById(id)) {
+            throw new ProductNotFoundException("Product not found with id: " + id);
+        }
         repo.deleteById(id);
     }
